training/train_detr.py
#!/usr/bin/env python3
"""
Training script for DETR Detection Model

End-to-end player and ball detection using Detection Transformer.

Usage:
    python train_detr.py --config configs/detr.yaml
"""

import argparse
import yaml
from loguru import logger

# TODO: Implement DETR training
# - Load detection annotations
# - Train transformer-based detector
# - Use Hungarian matching loss
# - Evaluate mAP metrics

def main():
    parser = argparse.ArgumentParser()
    parser.add_argument("--config", type=str, required=True)
    args = parser.parse_args()

    with open(args.config) as f:
        config = yaml.safe_load(f)

    logger.info("DETR training - Implementation placeholder")
    logger.info(f"Config: {config['model']['name']}")

if __name__ == "__main__":
    main()
