requirements.txt
# Deep Learning & Computer Vision
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.2.0
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.11.0

# Tracking
supervision>=0.22.0
lapx>=0.5.0

# Homography & Geometry
kornia>=0.7.0
scikit-image>=0.21.0

# Re-identification
timm>=1.0.0

# Jersey OCR
paddlepaddle>=2.5.0
paddleocr>=2.7.0

# Visualization
matplotlib>=3.7.0
mplsoccer>=1.2.0
plotly>=5.15.0

# Data Processing
pandas>=2.0.0
polars>=0.20.0
pyyaml>=6.0

# Utilities
tqdm>=4.65.0
loguru>=0.7.0
typer>=0.9.0
rich>=13.5.0
decord>=0.6.0

# ML utilities
scikit-learn>=1.3.0

# Training Infrastructure
albumentations>=1.3.0
wandb>=0.15.0
tensorboard>=2.13.0
SoccerNet>=0.1.7

# Graph Neural Networks
torch-geometric>=2.4.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0
torch-cluster>=1.6.0

# Real-time Optimization (Optional)
# Uncomment for TensorRT acceleration (requires CUDA):
# tensorrt>=8.6.0
# pycuda>=2022.1

# Uncomment for ONNX Runtime acceleration:
# onnx>=1.14.0
# onnxruntime-gpu>=1.15.0  # Use onnxruntime for CPU-only
